Applied shell wrapper to the current code coverage.
