+++ scripts/code_coverage_report/generate_code_coverage_report/Package.swift
@@ -43,6 +43,7 @@ let package = Package(
       name: "CoverageReportGenerator",
       dependencies: [
         .product(name: "ArgumentParser", package: "swift-argument-parser"),
+        "Utils"
       ]
     ),
     .target(
@@ -51,5 +52,8 @@ let package = Package(
         .product(name: "ArgumentParser", package: "swift-argument-parser"),
       ]
     ),
+    .target(
+      name: "Utils"
+    )
   ]
 )
